add container specific remote methods if a model is connected to loopback-component-storage ds
